Move llm prompt modal to own file
