+++ tater/ui/upload_layout.py
@@ -25,212 +25,11 @@ from tater.ui.schema_builder import (
     build_schema_builder_modal,
     register_schema_builder_callbacks,
 )
+from tater.ui.llm_prompt import (
+    build_llm_prompt_modal,
+    register_llm_prompt_callbacks,
+)
 
-
-# ---------------------------------------------------------------------------
-# LLM schema-design prompt
-# ---------------------------------------------------------------------------
-
-_LLM_PROMPT = """\
-You are a schema design assistant for Tater, a document annotation tool.
-
-Your goal: help the user design a Tater JSON annotation schema through \
-conversation, then produce valid JSON they can paste into the app.
-
-## How to proceed
-
-1. Ask: What kind of documents are you annotating? What information do you \
-need to capture about each one?
-2. For each piece of information, determine the right field type (see \
-reference below).
-3. Clarify options for choice/multi-choice fields, and entity type names for \
-span annotation fields.
-4. Ask whether any fields should only appear when another field has a specific \
-value (conditional visibility).
-5. Ask which fields are required (shown with a * indicator and tracked for \
-completion).
-6. Generate the final JSON schema in a code block.
-7. Offer to revise based on feedback.
-
-## Tater JSON schema format
-
-{
-  "spec_version": "1.0",
-  "title": "Your Schema Title",
-  "data_schema": [ /* field definitions */ ]
-}
-
-## Field types and valid widget types
-
-Field type     | Use for                           | Valid widget types (default first)
----------------|-----------------------------------|--------------------------------------------
-choice         | Single selection from a list      | segmented_control, radio_group, select, chip_radio
-multi_choice   | Multiple selections from a list   | multi_select, checkbox_group
-boolean        | True / false, yes / no            | checkbox, switch, chip_boolean
-text           | Free-form text                    | text_input, text_area
-numeric        | A number                          | number_input, slider
-range_slider   | A numeric range (low, high)       | range_slider (only option)
-span_annotation| Highlight spans of text           | span_annotation, span_popup
-divider        | Visual section break (no data)    | divider (only option)
-
-## Field definition format
-
-Standard field:
-{
-  "id": "field_id",              // unique, snake_case identifier
-  "type": "choice",              // field type from table above
-  "options": ["a", "b", "c"],    // required for choice and multi_choice only
-  "widget": {
-    "type": "segmented_control", // widget type — must match field type
-    "label": "Display Label",
-    "description": "Optional help text shown below the widget",
-    "required": true,            // optional; shows * and tracks completion
-    "placeholder": "hint...",    // text_input and text_area only
-    "min_value": 0,              // number_input, slider, range_slider
-    "max_value": 10,
-    "step": 1,
-    "entity_types": ["Person"],  // span_annotation only (omit for unlabeled spans)
-    "conditional_on": {"field": "other_field_id", "value": "some_value"}
-  }
-}
-
-Divider (section separator — no data field):
-{"widget": {"type": "divider", "label": "Section Heading"}}
-
-## Rules
-
-- "id" values must be unique and snake_case.
-- choice and multi_choice require "options" with at least 2 strings.
-- span_annotation with entity_types produces labeled spans; omit or leave \
-empty for unlabeled.
-- conditional_on hides the widget until the named field equals the given value.
-- "required" is UI-only — shows * and tracks progress; does not prevent saving.
-
-## Complete example
-
-{
-  "spec_version": "1.0",
-  "title": "Clinical Note Review",
-  "data_schema": [
-    {
-      "id": "relevance",
-      "type": "choice",
-      "options": ["relevant", "not_relevant", "unclear"],
-      "widget": {
-        "type": "segmented_control",
-        "label": "Relevance",
-        "required": true
-      }
-    },
-    {
-      "id": "confidence",
-      "type": "choice",
-      "options": ["high", "medium", "low"],
-      "widget": {
-        "type": "radio_group",
-        "label": "Confidence",
-        "conditional_on": {"field": "relevance", "value": "relevant"}
-      }
-    },
-    {
-      "id": "notes",
-      "type": "text",
-      "widget": {
-        "type": "text_area",
-        "label": "Notes",
-        "placeholder": "Any additional observations"
-      }
-    },
-    {"widget": {"type": "divider", "label": "Span Annotation"}},
-    {
-      "id": "conditions",
-      "type": "span_annotation",
-      "widget": {
-        "type": "span_annotation",
-        "label": "Medical Conditions",
-        "entity_types": ["Diagnosis", "Symptom", "Medication"]
-      }
-    }
-  ]
-}
-
----
-
-Begin by asking the user what they are annotating and what information they \
-need to capture.
-"""
-
-
-def _build_llm_prompt_modal() -> dmc.Modal:
-    return dmc.Modal(
-        id="llm-prompt-modal",
-        title=dmc.Group(
-            [
-                DashIconify(icon="tabler:robot", width=18),
-                dmc.Text("Design with AI", fw=600),
-            ],
-            gap="xs",
-        ),
-        opened=False,
-        size="lg",
-        closeOnEscape=True,
-        children=[
-            dmc.Text(
-                "Copy the prompt below and paste it into an AI assistant "
-                "(e.g. Claude, ChatGPT). It will guide you through designing "
-                "a schema and produce JSON you can paste back here.",
-                size="sm",
-                c="dimmed",
-            ),
-            dmc.Group(
-                [
-                    dmc.Text("Step 1 — copy the prompt", size="sm", fw=500),
-                    dmc.CopyButton(
-                        value=_LLM_PROMPT,
-                        children=DashIconify(icon="fa-regular:copy"),
-                        copiedChildren=DashIconify(icon="fa-regular:check-circle"),
-                        color="gray",
-                        copiedColor="dark",
-                        variant="transparent",
-                        px="xs",
-                    ),
-                ],
-                gap="xs",
-                align="center",
-                mt="md",
-            ),
-            dmc.Divider(mt="xs", mb="md"),
-            dmc.Text("Step 2 — paste the prompt into an AI assistant and follow its instructions", size="sm", fw=500),
-            dmc.Divider(mt="xs", mb="md"),
-            dmc.Text("Step 3 — paste the resulting schema below", size="sm", fw=500, mb="xs"),
-            dmc.Textarea(
-                id="llm-prompt-schema-input",
-                placeholder='{\n  "spec_version": "1.0",\n  "title": "...",\n  "data_schema": [...]\n}',
-                autosize=True,
-                minRows=6,
-                maxRows=16,
-                styles={"input": {"fontFamily": "monospace", "fontSize": "12px"}},
-            ),
-            html.Div(id="llm-prompt-feedback", style={"minHeight": "20px", "marginTop": "4px"}),
-            dmc.Group(
-                [
-                    dmc.Button(
-                        "Cancel",
-                        id="llm-prompt-close-btn",
-                        variant="subtle",
-                        color="gray",
-                    ),
-                    dmc.Button(
-                        "Apply schema",
-                        id="llm-prompt-apply-btn",
-                        n_clicks=0,
-                    ),
-                ],
-                justify="flex-end",
-                mt="md",
-            ),
-        ],
-    )
 
 
 # ---------------------------------------------------------------------------
@@ -509,7 +308,7 @@ def build_upload_layout(show_disclaimer: bool = False) -> dmc.MantineProvider:
         children=[
             dcc.Location(id="upload-location", refresh=True),
             build_schema_builder_modal(),
-            _build_llm_prompt_modal(),
+            build_llm_prompt_modal(),
             *(
                 [dmc.Modal(
                     id="disclaimer-modal",
@@ -992,51 +791,7 @@ def register_upload_callbacks(app: Dash, on_session_ready=None, show_disclaimer:
             return no_update, _error_text(f"Error loading example '{example_name}': {e}")
 
     register_schema_builder_callbacks(app)
-
-    @app.callback(
-        Output("llm-prompt-modal", "opened"),
-        Input("llm-prompt-open-btn", "n_clicks"),
-        Input("llm-prompt-close-btn", "n_clicks"),
-        prevent_initial_call=True,
-    )
-    def toggle_llm_prompt_modal(open_clicks, close_clicks):
-        if ctx.triggered_id == "llm-prompt-open-btn":
-            return True
-        return False
-
-    @app.callback(
-        Output("schema-store", "data", allow_duplicate=True),
-        Output("schema-feedback", "children", allow_duplicate=True),
-        Output("pending-hierarchies", "data", allow_duplicate=True),
-        Output("hierarchy-files-store", "data", allow_duplicate=True),
-        Output("llm-prompt-modal", "opened", allow_duplicate=True),
-        Output("llm-prompt-feedback", "children"),
-        Input("llm-prompt-apply-btn", "n_clicks"),
-        State("llm-prompt-schema-input", "value"),
-        prevent_initial_call=True,
-    )
-    def apply_llm_schema(n_clicks, text):
-        if not n_clicks or not (text or "").strip():
-            return no_update, no_update, no_update, no_update, no_update, no_update
-        try:
-            data = json.loads(text)
-        except Exception as e:
-            return no_update, no_update, no_update, no_update, no_update, _error_text(f"Invalid JSON: {e}")
-        ok, msg = _validate_schema_json(data)
-        if not ok:
-            return no_update, no_update, no_update, no_update, no_update, _error_text(msg)
-        pending = {
-            name: Path(source).name
-            for name, source in data.get("hierarchies", {}).items()
-            if isinstance(source, str)
-        }
-        field_names = [
-            f.get("id") for f in data.get("data_schema", [])
-            if f.get("id") and f.get("type") != "divider"
-        ]
-        n = len(field_names)
-        summary = _success_text(f"Schema from AI — {n} top-level field(s)")
-        return data, summary, pending, {}, False, None
+    register_llm_prompt_callbacks(app)
 
 
 # ---------------------------------------------------------------------------
